+++ src/Async/Await.ts
@@ -1,5 +1,0 @@
-interface Props {
-  children: any;
-}
-
-export const Await = ({ children }: Props) => children;