+++ src/Async/Catch.ts
@@ -1,5 +1,0 @@
-interface Props {
-  children: any;
-}
-
-export const Catch = ({ children }: Props) => children;