+++ src/Async/Finally.ts
@@ -1,5 +1,0 @@
-interface Props {
-  children: any;
-}
-
-export const Finally = ({ children }: Props) => children;