+++ src/Async/Try.ts
@@ -1,5 +1,0 @@
-interface Props {
-  children: any;
-}
-
-export const Try = ({ children }: Props) => children;