+++ src/Condition/Else.ts
@@ -1,5 +1,0 @@
-interface Props {
-  children: any;
-}
-
-export const Else = ({ children }: Props) => children;