+++ src/Condition/Then.ts
@@ -1,5 +1,0 @@
-interface Props {
-  children: any;
-}
-
-export const Then = ({ children }: Props) => children;