+++ src/Condition/Unless.ts
@@ -1,8 +1,0 @@
-interface Props {
-  children: any;
-  condition: boolean;
-}
-
-export const Unless = ({ children, condition }: Props) => {
-  return condition ? null : children;
-};