+++ src/Condition/When.ts
@@ -1,8 +1,0 @@
-interface Props {
-  children: any;
-  condition: boolean;
-}
-
-export const When = ({ children, condition }: Props) => {
-  return condition ? children : null;
-};