+++ src/Loop/Continue.ts
@@ -1,1 +1,0 @@
-export const Continue = () => null;