+++ src/Promise/Reject.ts
@@ -1,1 +1,0 @@
-export const Reject = ({ children }: any) => children;