+++ src/Promise/Resolve.ts
@@ -1,5 +1,0 @@
-interface Props {
-  children: any;
-}
-
-export const Resolve = ({ children }: Props) => children;