+++ src/Switch/Case.ts
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const Case: React.FC = ({ children }) => children;