+++ stories/Functions.stories.tsx
@@ -7,23 +7,22 @@ export default {
 };
 
 export const map = () => (
-  <Map data={[1, 2, 3]}>{(item: any) => <p>{item}</p>}</Map>
+  <Map data={[1, 2, 3]}>{(item: number[]) => <p>{item}</p>}</Map>
 );
 
 export const filter = () => (
-  <Filter
-    data={[1, 2, 3]}
-    // data={array('Data', [1, 2, 3])}
-    pattern={(item: any) => item !== 1}
-  >
-    {(filtered: any) => (
-      <Map data={filtered}>{(item: any) => <p>{item}</p>}</Map>
+  <Filter data={[1, 2, 3]} pattern={(item: number) => item !== 1}>
+    {(filtered: number[]) => (
+      <Map data={filtered}>{(item: number) => <p>{item}</p>}</Map>
     )}
   </Filter>
 );
 
 export const reduce = () => (
-  <Reduce data={[1, 2, 3]} pattern={(prev: any, next: any) => prev + next}>
-    {(reduced: any) => reduced}
+  <Reduce
+    data={[1, 2, 3]}
+    pattern={(prev: number, next: number) => prev + next}
+  >
+    {(reduced: number) => reduced}
   </Reduce>
 );