+++ CHANGELOG.md
@@ -8,11 +8,14 @@
 - loops: `For`, `ForOf`, `ForIn`, `ForEach`, `While`, `DoWhile`, `Break`, `Continue`
 - promises: `Promise`, `Resolve`, `Reject`
 
+## 1.0.0 beta 1 (Soon)
+
+- conditions: `If`, `Else`, `Then`, `When`, `Unless`
+- switch: `Switch`, `Case`, `Default`
+
 ## 1.0.0 alpha 1 / 1.0.0 alpha 2 (2020-02-18)
 
 #### New Feature
 
-- conditions: `If`, `Else`, `Then`, `When`, `Unless`
-- switch: `Switch`, `Case`, `Default`
 - timeouts: `SetTimeout`, `SetInterval`
 - functions: `Map`, `Filter`, `Reduce`